move ingestion to its own crate
