+++ crates/sui-indexer-common/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "sui-indexer-common"
+version.workspace = true
+authors = ["Mysten Labs <build@mystenlabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+axum.workspace = true
+backoff.workspace = true
+tokio.workspace = true
+tokio-util.workspace = true
+tokio-stream.workspace = true
+tracing.workspace = true
+telemetry-subscribers.workspace = true
+url.workspace = true
+clap.workspace = true
+reqwest.workspace = true
+futures.workspace = true
+prometheus.workspace = true
+thiserror.workspace = true
+
+sui-types.workspace = true
+sui-storage.workspace = true
+mysten-metrics.workspace = true
+
+
+[dev-dependencies]
+rand.workspace = true
+tempfile.workspace = true
+wiremock.workspace = true
