+++ crates/sui-ingestion/src/error.rs
