+++ crates/sui-ingestion/src/utils.rs
@@ -1,0 +1,124 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use futures::{Stream, StreamExt};
+use std::future::Future;
+use std::panic;
+use std::pin::pin;
+use tokio::task::JoinSet;
+
+/// Extension trait introducing `try_for_each_spawned` to all streams.
+pub trait TrySpawnStreamExt: Stream {
+    /// Attempts to run this stream to completion, executing the provided asynchronous closure on
+    /// each element from the stream as elements become available.
+    ///
+    /// This is similar to [StreamExt::for_each_concurrent], but it may take advantage of any
+    /// parallelism available in the underlying runtime, because each unit of work is spawned as
+    /// its own tokio task.
+    ///
+    /// The first argument is an optional limit on the number of tasks to spawn concurrently.
+    /// Values of `0` and `None` are interpreted as no limit, and any other value will result in no
+    /// more than that many tasks being spawned at one time.
+    ///
+    /// ## Safety
+    ///
+    /// This function will panic if any of its futures panics, will return early with success if
+    /// the runtime it is running on is cancelled, and will return early with an error propagated
+    /// from any worker that produces an error.
+    fn try_for_each_spawned<Fut, F, E>(
+        self,
+        limit: impl Into<Option<usize>>,
+        f: F,
+    ) -> impl Future<Output = Result<(), E>>
+    where
+        Fut: Future<Output = Result<(), E>> + Send + 'static,
+        F: FnMut(Self::Item) -> Fut,
+        E: Send + 'static;
+}
+
+impl<S: Stream + Sized + 'static> TrySpawnStreamExt for S {
+    async fn try_for_each_spawned<Fut, F, E>(
+        self,
+        limit: impl Into<Option<usize>>,
+        mut f: F,
+    ) -> Result<(), E>
+    where
+        Fut: Future<Output = Result<(), E>> + Send + 'static,
+        F: FnMut(Self::Item) -> Fut,
+        E: Send + 'static,
+    {
+        // Maximum number of tasks to spawn concurrently.
+        let limit = match limit.into() {
+            Some(0) | None => usize::MAX,
+            Some(n) => n,
+        };
+
+        // Number of permits to spawn tasks left.
+        let mut permits = limit;
+        // Handles for already spawned tasks.
+        let mut join_set = JoinSet::new();
+        // Whether the worker pool has stopped accepting new items and is draining.
+        let mut draining = false;
+        // Error that occurred in one of the workers, to be propagated to the called on exit.
+        let mut error = None;
+
+        let mut self_ = pin!(self);
+
+        loop {
+            tokio::select! {
+                next = self_.next(), if !draining && permits > 0 => {
+                    if let Some(item) = next {
+                        permits -= 1;
+                        join_set.spawn(f(item));
+                    } else {
+                        // If the stream is empty, signal that the worker pool is going to
+                        // start draining now, so that once we get all our permits back, we
+                        // know we can wind down the pool.
+                        draining = true;
+                    }
+                }
+
+                Some(res) = join_set.join_next() => {
+                    match res {
+                        Ok(Err(e)) if error.is_none() => {
+                            error = Some(e);
+                            permits += 1;
+                            draining = true;
+                        }
+
+                        Ok(_) => permits += 1,
+
+                        // Worker panicked, propagate the panic.
+                        Err(e) if e.is_panic() => {
+                            panic::resume_unwind(e.into_panic())
+                        }
+
+                        // Worker was cancelled -- this can only happen if its join handle was
+                        // cancelled (not possible because that was created in this function),
+                        // or the runtime it was running in was wound down, in which case,
+                        // prepare the worker pool to drain.
+                        Err(e) => {
+                            assert!(e.is_cancelled());
+                            permits += 1;
+                            draining = true;
+                        }
+                    }
+                }
+
+                else => {
+                    // Not accepting any more items from the stream, and all our workers are
+                    // idle, so we stop.
+                    if permits == limit && draining {
+                        break;
+                    }
+                }
+            }
+        }
+
+        if let Some(e) = error {
+            Err(e)
+        } else {
+            Ok(())
+        }
+    }
+}
